python/requirements.txt
azure.identity
azure.mgmt.dns
netifaces